backend/pyproject.toml
[tool.poetry]
name = "tarot-backend"
version = "0.1.0"
description = "Backend for Mystic Tarot NFT application"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.1"
Pillow = "^9.0.0"
stability-sdk = "^0.8.0"
python-dotenv = "^0.19.0"
web3 = "^6.0.0"
jinja2 = "^3.0.1"
requests = "^2.31.0"
flask-cors = "^4.0.0"
pyjwt = "^2.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
pytest-asyncio = "^0.21.1"
pytest-env = "^1.1.1"
coverage = "^7.3.2"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError"
]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.poetry.scripts]
start = "tarot.app:main"
